+++ src/models/Settings.php
@@ -57,6 +57,17 @@ class Settings extends Model
      * @var bool
      */
     public $excludeTemplates = 'cpresources,';
+
+    // Config FIles
+    /**
+     * @var bool
+     */
+    public $enableConfigFiles = false;
+
+    /**
+     * @var bool
+     */
+    public $excludeConfigFiles = 'cpresources,';
     // Logs
     /**
      * @var bool
@@ -296,13 +307,19 @@ class Settings extends Model
     public $notificationReplyToEmail;
 
     /**
+     * @var integer
+     */
+    public $maxExecutionTime = 3600;
+
+    /**
      * @inheritdoc
      */
     public function rules()
     {
         return [
             ['backupsAmount', 'integer', 'min' => 1, 'on' => 'general'],
-            ['backupsAmount', 'required', 'on' => 'general'],
+            ['maxExecutionTime', 'integer', 'min' => 300, 'on' => 'general'],
+            [['backupsAmount', 'maxExecutionTime'], 'required', 'on' => 'general'],
             [
                 ['enableDatabase', 'enableTemplates', 'enableLocalVolumes', 'enableLogs'],
                 BackupFilesValidator::class, 'on' => 'backupFiles'